requirements.txt
Flask==3.1.*
Flask-SQLAlchemy==3.1.*
Flask-Login
Flask-WTF
Flask-Bcrypt
SQLAlchemy==2.0.*
PyMySQL
python-dotenv
bcrypt
email_validator